Move util function to become Sensor attribute
